modified textcat to make it usable with its own training data

+++ src/hydrat/wrapper/textcat.py
@@ -11,16 +11,22 @@ import operator
 from subprocess import Popen, PIPE, STDOUT
 
 class TextCat(object):
-  def __init__(self, toolpath, scratch):
-    self.model_path = None
+  def __init__(self, toolpath, scratch='/tmp', modelpath=None):
     self.toolpath = toolpath
     self.scratch = scratch
+    self.model_path = modelpath
+    self.trained = False
 
   def __del__(self):
-    if self.model_path is not None:
+    # Must check if we trained this model to avoid deleting external models
+    if self.trained and self.model_path is not None:
       shutil.rmtree(self.model_path)
 
   def train(self, pairs):
+    # Delete old trained model if any
+    if self.trained:
+      shutil.rmtree(self.model_path)
+      
     # Create a temporary directory to store models
     self.model_path = tempfile.mkdtemp(prefix='textcat', dir=self.scratch)
 
@@ -30,6 +36,9 @@ class TextCat(object):
       class_data = '\n'.join(t for t,c in group)
       p = Popen([self.toolpath, '-n'], stdout=open(class_path,'w'), stdin=PIPE, stderr=STDOUT)
       p.communicate(input=class_data)[0]
+
+    # Mark as trained
+    self.trained = True
     
 
   def classify(self, text):
